JUNK-step4_backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
requests==2.31.0
sqlalchemy==2.0.23
pandas==2.1.4
numpy==1.24.3
graphene==3.3
python-dotenv==1.0.0
pymysql==1.1.0
openai==1.3.8
httpx==0.25.0
psycopg2-binary==2.9.9
python-multipart==0.0.6
bcrypt==4.1.2
pydantic==2.5.2
python-jose[cryptography]==3.3.0